Split editor code
